requirements.txt
dill==0.3.7
Flask==3.0.0
joblib==1.3.2
matplotlib==3.6.3
numpy==1.23.5
pandas==1.5.3
plotly==5.18.0
PyYAML==6.0.1
Requests==2.31.0
shap==0.43.0
streamlit==1.28.0
streamlit_authenticator==0.2.3
pycaret==3.0.0